+++ clients/admin-ui/src/app/store.ts
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import { setupListeners } from "@reduxjs/toolkit/query/react";
 import { createWrapper } from "next-redux-wrapper";
 
-import { datasetApi } from "~/features/dataset";
+import { datasetApi, reducer as datasetReducer } from "~/features/dataset";
 import { systemApi } from "~/features/system";
 import { reducer as userReducer } from "~/features/user";
 
@@ -10,6 +10,7 @@ const makeStore = () => {
   const store = configureStore({
     reducer: {
       user: userReducer,
+      dataset: datasetReducer,
       [datasetApi.reducerPath]: datasetApi.reducer,
       [systemApi.reducerPath]: systemApi.reducer,
     },